Bug fixing, add header files for every file without one, adjusts data types, now compiles properly
Movement.c, fruit_generation.c and mainmenu.c now have their equivalent .h header files, which are what is included in snake.c
Several functions have had their data types adjusted (eg some functions were still using a signed char for the playspace).
Added the -lm compilation flag to makefile, sqrt() and pow() now working as intended.
The make command now runs without erros.

+++ drawing.h
@@ -10,12 +10,12 @@ The color parameter decides which color to use,
 
 #include <stdint.h>
 //Draws a square of white, size depends on scal
-void draw_square(int x, int y, int scale, unsigned short** buffer, int color);
+void draw_square(int x, int y, int scale, unsigned short** buffer, unsigned short color);
 
 //Draws a vertical line using the draw square function
-void draw_vert_line(int start_x, int start_y, int length, int scale, unsigned short** buffer, int color);
+void draw_vert_line(int start_x, int start_y, int length, int scale, unsigned short** buffer, unsigned short color);
 
 //Draws a horizontal line using the draw square function
-void draw_horziontal_line(int start_x, int start_y, int length, int scale, unsigned short** buffer, int color);
+void draw_horziontal_line(int start_x, int start_y, int length, int scale, unsigned short** buffer, unsigned short color);
 
 #endif  /*DRAWING_H*/
